fix claim surface area overflowing int on huge regions
getSurfaceArea() multiplied two ints, so a region near the +/-32,000,000
coordinate limit wrapped around to a small or negative number. that let
createClaimAt/resizeClaim's claim-block cost checks pass for free since
"blocks < surfaceArea" is never true against a negative area, letting
someone claim or extend a claim to a huge size without the blocks to
pay for it. do the math in long instead.

+++ common/src/main/java/net/william278/huskclaims/claim/ClaimEditor.java
@@ -347,7 +347,7 @@ public interface ClaimEditor {
             // Send the correct deletion message
             if (claim.getOwner().isPresent()) {
                 getPlugin().getLocales().getLocale("claim_deleted",
-                                Integer.toString(claim.getRegion().getSurfaceArea()))
+                                Long.toString(claim.getRegion().getSurfaceArea()))
                         .ifPresent(executor::sendMessage);
                 return;
             }
